Read library name from own base
Before this patch, PatchWork assumed that the library base is loaded
into the A6 register. On a hit, the library name was read based from the
contents found there.

However, although it is required to load the library base into A6 before
invoking a function, there is no guaratee that it is actually done.

With this patch, PatchWork now uses an own local library base for
reading the library name for a hit output.